Add typecast to a test
(1 << x) where "x" is a uint64_t would treat "1" as an int and
roll over at 32-bits. Typecasting "1" to uint64_t eliminates
that and supports a full 64-bit value.

+++ testsuite/expect/test1.91.prog.c
@@ -100,7 +100,7 @@ uint64_t mask_to_int(cpu_set_t *mask)
 				rc = 999999999;
 				break;
 			}
-			rc += (1 << i);
+			rc += (((uint64_t) 1) << i);
 		}
 	}
 	return rc;
